+++ src/krux/pages/qr_view.py
@@ -384,6 +384,45 @@ class SeedQRView(Page):
         bmp_img.save("/sd/" + file_name)
         self.flash_text(t("Saved to SD card") + ":\n%s" % file_name)
 
+    def save_svg_image(self, file_name):
+        """Save QR code image as .svg file"""
+        from ..sd_card import SVG_IMAGE_EXTENSION
+        from .file_operations import SaveFile
+
+        import image
+        import lcd
+
+        scale = 10
+
+        code, size = self.add_frame(self.code, self.qr_size)
+
+        save_page = SaveFile(self.ctx)
+        file_name = save_page.set_filename(
+            file_name, file_extension=SVG_IMAGE_EXTENSION
+        )
+        if file_name == ESC_KEY:
+            return
+
+        self.ctx.display.clear()
+        self.ctx.display.draw_centered_text(t("Processing.."))
+
+        with open("/sd/" + file_name, 'w') as f:
+            width = size * scale
+            height = width
+            f.write('<svg xmlns="http://www.w3.org/2000/svg" width="{}" height="{}">'.format(width,height))
+            for y_index in range(0, size):
+                for x_index in range(0, size):
+                    index = y_index * size + x_index
+                    bit_value = (code[index >> 3] >> (index % 8)) & 1
+                    if bit_value:
+                        x = x_index*scale
+                        y = y_index*scale
+                        f.write('<rect stroke="black" stroke-width="0" x="{}" y="{}" width="{}" height="{}" fill="black"/>'.format(x,y,scale,scale))
+            f.write('</svg>')
+            f.close()
+
+        self.flash_text(t("Saved to SD card") + ":\n%s" % file_name)
+
     def save_qr_image_menu(self):
         """Options to save QR codes as images on SD card"""
 
@@ -419,6 +458,12 @@ class SeedQRView(Page):
                     ),
                 )
             )
+        qr_menu.append(
+            (
+                "SVG",
+                lambda: self.save_svg_image(suggested_file_name),
+            )
+        )
         submenu = Menu(self.ctx, qr_menu, offset=2 * FONT_HEIGHT, back_label=None)
         submenu.run_loop()
         return MENU_CONTINUE